inline kernel steps
