+++ honeycomb-kernels/src/grisubal/kernel/mod.rs
@@ -60,106 +60,6 @@ macro_rules! up_intersec {
     }};
 }
 
-/// Inner building routine.
-///
-/// This function builds a combinatorial map from the described geometry. The returned
-/// map is an adjusted grid that can be clipped in order to keep only part of the mesh.
-/// See [`grisubal::Clip`] for more information.
-///
-/// # Arguments
-///
-/// - `geometry: &Geometry2<T>` -- Description of the input geometry.
-///
-/// ## Generics
-///
-/// - `T: CoordsFloat` -- Floating point type used for coordinate representation.
-pub fn build_mesh<T: CoordsFloat>(
-    geometry: &Geometry2<T>,
-    [cx, cy]: [T; 2],
-    [nx, ny]: [usize; 2],
-    origin: Vertex2<T>,
-) -> CMap2<T> {
-    #[cfg(feature = "profiling")]
-    let mut instant = std::time::Instant::now();
-
-    // compute grid characteristics
-    // build grid descriptor
-    let ogrid = GridDescriptor::default()
-        .n_cells_x(nx)
-        .n_cells_y(ny)
-        .len_per_cell_x(cx)
-        .len_per_cell_y(cy)
-        .origin(origin);
-
-    // build initial map
-    let mut cmap = CMapBuilder::default()
-        .grid_descriptor(ogrid)
-        .add_attribute::<Boundary>() // will be used for clipping
-        .build()
-        .expect("E: unreachable"); // urneachable because grid dims are valid
-
-    #[cfg(feature = "profiling")]
-    unsafe_time_section!(instant, Section::BuildMeshInit);
-
-    // process the geometry
-
-    // STEP 1
-    // the aim of this step is to build an exhaustive list of the segments making up
-    // the GEOMETRY INTERSECTED WITH THE GRID, i.e. for each segment, if both vertices
-    // do not belong to the same cell, we break it into sub-segments until it is the case.
-
-    let (new_segments, intersection_metadata) =
-        generate_intersection_data(&cmap, geometry, [nx, ny], [cx, cy], origin);
-
-    #[cfg(feature = "profiling")]
-    unsafe_time_section!(instant, Section::BuildMeshIntersecData);
-
-    // STEP 1.5
-    // precompute stuff to
-    // - parallelize step 2
-    // - make step 2 and step 3 independent from each other
-
-    let n_intersec = intersection_metadata.len();
-    let (edge_intersec, dart_slices) =
-        group_intersections_per_edge(&mut cmap, intersection_metadata);
-    let intersection_darts = compute_intersection_ids(n_intersec, &edge_intersec, &dart_slices);
-
-    // STEP 2
-    // insert the intersection vertices into the map & recover their encoding dart. The output Vec has consistent
-    // indexing with the input Vec, meaning that indices in GeometryVertex::Intersec instances are still valid.
-
-    insert_intersections(&mut cmap, &edge_intersec, &dart_slices);
-
-    #[cfg(feature = "profiling")]
-    unsafe_time_section!(instant, Section::BuildMeshInsertIntersec);
-
-    // STEP 3
-    // now that we have a list of "atomic" (non-dividable) segments, we can use it to build the actual segments that
-    // will be inserted into the map. Intersections serve as anchor points for the new segments while PoI make up
-    // "intermediate" points of segments.
-
-    let edges = generate_edge_data(&cmap, geometry, &new_segments, &intersection_darts);
-
-    #[cfg(feature = "profiling")]
-    unsafe_time_section!(instant, Section::BuildMeshEdgeData);
-
-    // STEP 4
-    // now that we have some segments that are directly defined between intersections, we can use some N-maps'
-    // properties to easily build the geometry into the map.
-    // This part relies heavily on "conventions"; the most important thing to note is that the darts in `MapEdge`
-    // instances are very precisely set, and can therefore be used to create all the new connectivities.
-
-    insert_edges_in_map(&mut cmap, &edges);
-
-    #[cfg(feature = "profiling")]
-    unsafe {
-        TIMERS[Section::BuildMeshInsertEdge as usize] = Some(instant.elapsed());
-    }
-
-    // return result
-    cmap
-}
-
 // --- type aliases for clarity
 
 pub type Segments = HashMap<GeometryVertex, GeometryVertex>;
